+++ crates/squawk_syntax/test_data/validation/select.sql
@@ -1,0 +1,7 @@
+
+from t select c;
+
+from t;
+
+with t as (from k)
+select * from t;